requirements.txt
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
xgboost==2.0.3
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
anthropic==0.18.1
APScheduler==3.10.4
